Solving Day6 Part1
First time using the Array2D.

I used an active pattern to match if the cell is within bounds, which I learned from Day3.

Learned that Array2D mutations are in place.

My first instinct was to map each symbol to an enum. However, I remembered (from when I used to do Leetcode) that I could denote a visited cell with an int, and keep adding to the counter every time it is visited.