Solving Day7 Part1
I kinda came up with this while on my sleep.

The idea of producing all possible permutations of operators with repetition. Then for each permutation zip the operator with the operand. Iterate until all operations are exhausted.

However, I lost so much time in the permutation. I just wanted to understand how it worked.

Current time complexity of the permutation generation is factorial because we are not memoizing sub-permutations.

With i32 the program overflows, so I had to use i64