Solving Day8 Part1
Again, I spent a lot of time trying to understand how to generate combination of elements.

Why combinations? If we think of each node as a point, or vector, then for a pair of vectors V1 and V2, the antinode is at V2 + (V2 - V1). Then we just have to check whether the antinode is in the grid.

This was my first time using the Vector2 struct. I first though on creating my own, but ultimately strayed away from it because it might have take me a lot of time.

On the downside, the Vector2 struct only support floats, so I had to convert to between int and float.

I am thinking to create my own iVector2 struct.

I like how my personal utils library is growing