Solving Day8 Part2
I am felling quite happy with this one.

I was able to extract the antinode generation function. In short we recursively check the next antinode using the displacement until we reach out of bounds.

I wonder how Vector2 does comparisons since their members are floats.

An int implementation of Vector2 that supports comparison or hashing so we can add it to a set, would have saved us searching for distinct entries.