Solving Day9 Part1
Not so hard.

Some highlights from this problem:

I added a function that reads a file and returns a stream of int using a seq computational expression.

I added a function that swaps in place two elements of an array given their indices

I like this solution because the only allocation happens during the compacting stage