Solving Day 10 Part 2
Okay, so it turns out that I actually had solved Part 2 when I first tried to solve Part 1. So it was quite easy.

However, I see there's some duplication between both solutions. The main difference is whether to use a Set or List to track the reached points.

I guess I could always use a List and use distinct? However, I kind of like the O(1) lookups a set provides.

Anyways, committing as it is for now