Solving Day11 Part02
My Part01 solution was not fit to solve Part02. It is too slow. 35 iterations would already take forever.

When I first implemented Part01 I thought of trying to solve the problem using dynamic programming, but didn't bother because I got the solution.

Anyways, the challenging part was to figure out how to memoize the data.

In this case using a dictionary instead of an immutable map, proved to be so much easier.